refactor(gate): 拆分 change_risk 巨石文件过 sentrux 棘轮 (#95)
Refs #95

sentrux_gate's god-file rule (sentrux_gate.rs:738, `loc > 800 || (functions > 25 && loc > 400)`) flagged the new change_risk.rs at 1157 lines; split into change_risk/{mod,git,signals,scoring,render,tests}.rs, each well under threshold, behavior/JSON/CLI/tests unchanged.
